+++ src/command_modules/azure-cli-vm/HISTORY.rst
@@ -10,6 +10,7 @@ Release History
 * `vm create`: support configure Public-IP SKU
 * `vm secret format`: Added extra validation. Added `--keyvault` and `--resource-group` to support scenarios
                       where the command is unable to resolve the vault ID. [#5718](https://github.com/Azure/azure-cli/issues/5718)
+* `vm/vmss create`: emit out a better error if resource group's location has no zone support
 
 2.0.29
 ++++++